requirements.txt
Django==4.1.7
requests==2.28.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2